pyproject.toml
[tool.poetry]
name = "school"
version = "0.1.0"
description = ""
authors = ["Kirill8769 <tkkrozn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.5.3"
drf-yasg = "^1.21.7"
stripe = "^9.10.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
